+++ go-server/service/math/vector.go
@@ -3,11 +3,11 @@ package math
 import "math"
 
 type Vector2 struct {
-	X float64
-	Y float64
+	X int
+	Y int
 }
 
-func NewVector2(x, y float64) Vector2 {
+func NewVector2(x, y int) Vector2 {
 	return Vector2{X: x, Y: y}
 }
 
@@ -16,7 +16,7 @@ func SubVec2(first, second Vector2) Vector2 {
 }
 
 func (v *Vector2) NormalizeVec() {
-	factor := math.Sqrt(math.Pow(v.X, 2) + math.Pow(v.Y, 2))
-	v.X = v.X / factor
-	v.Y = v.Y / factor
+	factor := math.Sqrt(math.Pow(float64(v.X), 2) + math.Pow(float64(v.Y), 2))
+	v.X = int(float64(v.X) / factor)
+	v.Y = int(float64(v.Y) / factor)
 }
